+++ tensorflow_io/core/python/api/experimental/__init__.py
@@ -14,8 +14,9 @@
 # ==============================================================================
 """tensorflow_io.experimental"""
 
-from tensorflow_io.core.python.experimental.io_tensor import IOTensor
 from tensorflow_io.core.python.experimental.io_dataset_ops import IODataset
+from tensorflow_io.core.python.experimental.io_tensor import IOTensor
+from tensorflow_io.core.python.experimental.io_layer import IOLayer
 
 from tensorflow_io.core.python.api.experimental import ffmpeg
 from tensorflow_io.core.python.api.experimental import image
