+++ tensorflow_io/core/python/api/v0/__init__.py
@@ -17,7 +17,6 @@
 # tensorflow_io.core.python.ops is implicitly imported (along with file system)
 from tensorflow_io.core.python.ops.io_dataset import IODataset
 from tensorflow_io.core.python.ops.io_tensor import IOTensor
-from tensorflow_io.core.python.ops.io_layer import IOLayer
 
 from tensorflow_io.core.python.api.v0 import genome
 from tensorflow_io.core.python.api.v0 import image
